Minor Codable conformance improvements in SwiftBuildSupport (#8449)
### Motivation:
Some of the `SwiftBuildSupport/PIFBuilder.swift` encoding/decoding code
was using unsafe string keys. This replaces with proper typesafe keys.

This part of the ground work to support the new PIF builder in
`SwiftBuildSupport` (i.e., rdar://147527170).

### Modifications:
When conforming to `Codable`, replace string keys with `enum` based keys
instead.

### Result:
The resulting code is a tad safer and easier to understand :-)

Tracked by rdar://148546582.