Tests: Enable more WorkspaceTests on Windows
Many WorkspaceTests instantiate a MockWorkspace, which takes a path as
an input. The tests were defining a POSIX-like path, whhich was not
supported on Windows.

Convert the test to use the `withTemporaryDirectory` fixture to specify
a path as an input to the MockWorkspace.  Most tests pass, however some
require adjsustments on the expectation.