Tests: Enable more WorkspaceTests on Windows
Many WorkspaceTests instantiate a MockWorkspace, which takes a path as
an input. The tests were defining a POSIX-like path as the sandbox
directory, however, some assertions were incorrect as, on these *nix like
filesystem, the canonical path name and the path were identical.

Update the WorkspaceTests expectation accordingly, and add some
automated tests to cover the InMemoryFileSystem.