SwiftSDK: Remove hardcoded WASI sysroot path derivation
This change removes the hardcoded path derivation for the WASI sysroot
from the `SwiftSDK.deriveTargetSwiftSDK` method. The previous implementation
assumed there is $TOOLCHAIN_ROOT/share/wasi-sysroot when `--triple
wasm32-unknown-wasi` is passed, but this is no longer the case with the
[deprecation of the Wasm toolchain
installation](https://github.com/swiftwasm/swift/issues/5604) in favor
of Swift SDKs.