Fix duplicate modulemap errors with macro and plugin deps
We were including flags to hook up modulemaps and include files
to C library dependencies in macros and plugin tools through to
the modules that depend on those. This adds the capability to
prune the depth first searches through the dependencies to ensure
these are skipped when crossing macro and plugin boundaries.