Fix bootstrapping on OpenBSD (#8451)
### Motivation:

Ensure swiftpm bootstraps successfully to fully build an OpenBSD
toolchain.

### Modifications:

* Add the nobtcfi linker flag to the bootstrap script for OpenBSD. This
is unconditional for now since swiftpm uses Concurrency liberally and
this triggers #80059, so swiftpm only builds on the configuration where
BTCFI is disabled. We can revisit some of that perhaps later.

* Ensure SPMSQLite3 builds with the correct link library search path
flag in CMakeLists.txt.

* Update Package.swift conditionals for SPMSQLite3.

### Result:

swiftpm successfully builds and bootstraps on OpenBSD.