Link test command test failures on Linux to issue (#8459)
The TestCommandTests fail on Linux due to a common linker problem with
swift build build system, which is the missing 'main' symbol problem. This
is described further detail in #8439.

Provide the link in the skip messages and remove the TODO tags since the
investigation is complete and the reason for the failures is now known.

Update the common test case logic so that it dumps the command execution
stdout and stderr in the case of failure to help diagnose these result.xml failures
immediately in the logs.