Add `swift test --attachments-path`. (#8492)
This PR adds support for the `--attachments-path` CLI argument on `swift
test` as approved in
[ST-0009](https://github.com/swiftlang/swift-evolution/blob/main/proposals/testing/0009-attachments.md).
We will maintain support for the older `--experimental-attachments-path`
version through Swift 6.2.

The implementation of this option is held entirely in Swift Testing at
this time, so no actual code is needed to support it, we just need to
make sure Swift Argument Parser doesn't complain about it.

Resolves rdar://147753584.

@plemarquand has integration tests but they're blocked on a newer
toolchain; unit testing of this argument exists in the Swift Testing
repo.