Make add-dependency idempotent
If a manifest already contains the exact dependency is being
added via a `add-dependency` call, don't add it twice. Instead, do
nothing. This ensures we don't end up with an invalid manifest with two
of the same dependencies in it.

If a manifest contains a dependency with the supplied URL but a
different version qualifier then throw an error explaining that the
supplied dependency has already been added.

Issue: #8519