[6.1] Fix prebuild commands (#8440) (#8490)
Somewhere along the way earling in the 6.1 branch changes were made to
how the prebuild commands did their check to make sure the executable
invoked in the command existed in the scratch dir.

It had always used a trick that kept track of whether a tool was a
vendored binaryTarget or built. And only succeed if it was vendored.
That information was lost in the restructuring.

This change stores whether a tool comes from a executable target (built
tool) or a binary target (vended) and properly checks to make sure the
tool is vended.

This was detected when the SwiftLint prebuild command target was failing
when a user upgraded to 6.1. This change will need to be cherry picked
back to the 6.1 branch for the next point release.