Multi-thread pulling of out put from AsyncProcess
We are seeing hangs on Windows when SwiftPM tests itself.
Looking at how AsyncProcess pulls stdout and stderr, I can see
that there may be conditions where the streams aren't fully drained
causing the hang. The proper way to do this, like it's done on
the other platforms is to create a thread for each stream.