Output Swift Build PIF JSON for Graphviz visualization (#8539)
### Motivation:

Improve visualization/debuggability of the PIF JSON sent over to
Swift Build (when using the new `--build-system swiftbuild` option).

### Modifications:

This PR introduces the `--print-pif-manifest-graph` option to write out
the PIF JSON sent to Swift Build as a Graphviz file. This follows a similar 
design we used for the existing `--print-manifest-job-graph` option.

All the serialization happens in the new `DotPIFSerializer.swift` 
file, with this function as the entry point:

  func writePIF(_ workspace: PIF.Workspace, toDOT outputStream: OutputByteStream)

### Result:

With Graphviz tool (https://graphviz.org) installed 
(i.e., provides the `dot` command below), we can then 
run commands like this:

$ swift build --build-system swiftbuild --print-pif-manifest-graph | dot -Tpng > PIF.png