Avoid building targets that are conditional on platform when building all (#8568)
When using the swiftbuild build system, and building all targets with a
project that contains a target that is depended on only when on a
specific platform that target should be skipped. Make the 
All(Excluding|Including)Tests generated targets depend only on targets
that are reachable from the roots through dependencies that are either
unconditional or are conditioned on the current platform.