Tests: Enable more WorkspaceTests on Windows (#8457)
Many WorkspaceTests instantiate a MockWorkspace, which takes a path as
an input. The tests were defining a POSIX-like path as the sandbox
directory, however, some assertions were incorrect as, on these *nix
like filesystem, the canonical path name and the path were identical.

Update the WorkspaceTests expectation accordingly, and add some
automated tests to cover the InMemoryFileSystem (which is commented out
as a crash occurs in the Linux Platform CI build with Swift Testing
tests).

Relates: #8433
rdar://148248105