Make add-dependency idempotent (#8534)
### Motivation:

If the same dependency is added twice with different version specifiers
the manifest is no longer valid. If the same dependency is added twice
with the same version specifier it remains valid, but less maintainable.

### Modifications:

If a manifest already contains the exact dependency and version that is
being added via a `add-dependency` call, don't add it twice. Instead, do
nothing.

If a manifest contains a dependency with the supplied URL but a
different version qualifier then throw an error explaining that the
supplied dependency has already been added.

### Result:

A manifest that continues to parse correctly.

Issue: #8519