[Traits] Change `TraitConfiguration` to be an enum instead of struct (#8370)
Fixes #8355.

Since the way a `TraitConfiguration` is set up can affect how traits are
calculated and whether default traits are overridden or not, it's more
ergonomic to change this to an `enum` that lists the various possible
cases of configuration:

- No trait configuration, which defaults to using default traits if
applicable
- A configuration with an empty set of traits that overrides the default
traits
 - A configuration that enables all traits
- A configuration that specifies an exact set of traits that are enabled

This way we can avoid doing multiple checks on the properties of the
`struct` and instead rely on the exact case that describes how to
continue calculating enabled traits.

The refactoring done as a part of this change also fixes #8356.