Fix dup modmaps taking into account macro test targets. (#8524)
In my previous attempt to fix the duplicate modulemaps from C library
dependencies of macros/plugins and destination targets, I didn't take
into account test targets depending on macros, i.e. macro tests. In
these cases the modulemap flags do need to pass through to the tests.
This was discovered in swift-foundation which couldn't find the
_SwiftSyntaxCShims module map.

Fixed this my moving the traversing link dependencies a layer lower
since we need access to the parent node during traversal and this is
available in the successor methods. Cleaned up my previous attempt.