Forward cherry pick prebuilts feature to main.
To get the prebuilts feature out in time for 6.1, the bulk of the
work was done on the release/6.1 branch. Cherry pick those changes
to the main branch to get 6.2 caught up.

This is a manual cherry pick so it's going to be a bit hard to
follow. I did a diff between main and release/6.1 and copied all
the changes I made there.