Implement SwiftFixIt — a library for deserializing diagnostics and applying fix-its (#8585)
### Motivation:

This is part of the [migration
tooling](https://github.com/swiftlang/swift-evolution/pull/2673)
initiative for Swift features.

### Modifications:

For now, this API accepts a collection of `.dia` files, deserializes
them, and applies fix-its in place. Eventually, it is expected to serve
the upcoming `swift fixit` and `swift migrate` subcommands and support
various filtering options, as well as a stateful, interactive workflow
similar to `git add --patch`.