Use async methods when loading manifests (#8551)
### Motivation:

This follows on from my work to convert the registry code to `async`
methods.

### Modifications:

Switch to using async methods when loading manifests, converting
synchronous code to the simpler async equivalents and converting sync
API calls to their synchronous versions.

### Result:

This resolves several warnings and FIXMEs, makes the code easier to read
and generally implements concurrency in a more idiomatic way.