Remove Swift-numerics dependency
A `swift-numerics` package dependency was added to convert an XCTest
`XCTAssertEqual(_:_:accuracy:_:file:line:)` API do use the
`isApproximatelyEqual()` API in `swift-numerics`, but converting that
test resulted in a flaky behaviour.  As such, the test was not converted
to Swift Testing, but the swift-numerics dependency remained.

This change removes the swift-numerics dependency as it's no currently
required.