Clarify release branch cherry-picks process in `CONTRIBUTING.md` (#8634)
### Motivation:

SwiftPM's contributor documentation should clarify that in general changes should be submitted against the `main` branch first, with organization-wide PR description templates to be followed for release branch cherry-picks.

### Modifications:

Added a paragraph to `CONTRIBUTING.md` clarifying the process of the `swiftlang` GitHub org for release branches.

### Result:

Contributors documentation is more clear and it's more obvious how a change could be cherry-picked to release branches.