Migrate PackageDescription.md reference
The content for `Documentation/PackageDescription.md` is currently hosted
on https://docs.swift.org/package-manager/ and is considered legacy content
to be removed. The newer version of this documentation is hosted at
https://developer.apple.com/documentation/packagedescription for which the
content is found under `Sources/PackageDescription/PackageDescription.docc/PackageDescription.md`.

We'd like to direct users to this developer documentation from our DocC catalog, but since
DocC does not currently support external linking to other DocC catalogs we can
simply create an article in our catalog that provides a redirect link to above link.