Tests: Convert StaticBinaryLibrary Test to Swift Testing
When SwiftPM exeuctes tests using the native build system, the bundle
path was failing on MacOS as the `Bundle.allBundles` did not contain the
`.xctest` bundle.

If we fail to find the .xctest bundle, we inspect the command line
argument for the `--test-bundle-path` argument and contruct the macOS
bundle root based on the argument value.

In addition, convert the StaticBinaryLibraryTests to Swift Testing to
validate the logic works.