Detect embedded swift in flags
SwiftPM currently only detects Embedded Swift set via a swift settings
and uses this to enable/disable a couple features. This is nice, but
doesn't handle the case where Embedded is enabled via a user flag,
typically set in a toolset.json. This commit updates the logic to also
search for "-enable-experimental-feature Embedded" in the various place
user flags can be placed.