Async git repository opening
The `RepositoryProvider` protocol's `open` method is defined as
synchronous. Convert this to `async` to avoid concurrency workarounds in
clients that might be prone to deadlocks.

This involved porting `RepositoryManager.lookup` to structured
concurrency. The method has some concurrency constraints that are
preserved:

- There is a limit (`maxConcurrentOperations`) on how many simultaneous
  lookups can be performed concurrently.
- If a lookup is requested and one is in flight for the same
  `RepositorySpecifier`, the in flight request completes before the new
  one is started.

This PR also moves away from using a `DispatchQueue` to make delegate
calls, instead opting to use a `Task` that calls through an `actor`
based proxy to the underlying delegate. Gating these calls through an
`actor` allows us to ensure the calls are processed sequentially as they
were with the `DispatchQueue`.