Don't mark the compatibility header module map as requiring Objective-C
The module map generated for the compatibility header was marked as
`requires objc`. This triggers an error when importing it from a C source
file.

This declaration is superfluous, the compatibility header is already
printed in a way where the Objective-C code is protected behind a
language check. C clients can safely import the current compatibility
header even if it may not see any content.

Let's lift this restriction. It isn't currently necessary and we're
adding C content to the compatibility header with the official support
for `@cdecl` that is independent of Objective-C.