[Commands] Migrate: Group manifest update errors per target (#8677)
### Motivation:

If migrate fails to insert one or more features into the manifest for a
target, let's produce a single error that lists all of the features,
this makes the logging less noisy.

### Modifications:

- Update `SwiftPackageCommand` to request manifest to all of the
features related to a target and produce a single error if that fails.

### Result:

If manifest has some non-standard ways of defining targets or settings
that cannot be updated by `AddSettting`, migration logging is going to
be more organized and less noisy.