Dynamically determine the swift compiler version.
Prebuilts for macros and the automated downloading of SwiftSDKs
need to know the version of the compiler so we can fetch
compatible binaries. The swiftc compiler has a --print-target-info
options which dumps out a JSON structure that contains the
compilerVersion. We already use info in this structure to
determine the hostTriple for the UserToolchain.

This adds the swiftCompilerVersion to UserToolchain that uses
a couple of regex's to pull out the Swift compiler version.
This is then used by the prebuilts feature instead of our current
hardcodeing of the swift toolchain version.

This also turns the prebuilts feature on by default which was
supposed to be done in the last update.