Emit and search for the compatibility header in an `include` subdirectory
Emit the generated compatibility header and module map in a new
`include` directory under the temporary build directory of Swift
libraries. This directory can then be used as a header search path for
the dependent clang targets.

This replaces the previous header search strategy that used only the
generated module map. This was incompatible with C source files and
classic non-module textual includes. The use of a header search path
should support both module imports via the module map as well as
includes of the compatibility header.