6.2: Use WasmKit by default for Wasm triples in toolsets (#8692)
Cherry-pick of #8668, merged as 02fe923fcd410a7b2f04967a85cde18e98f1d820

**Explanation**: As [WasmKit](https://github.com/swiftwasm/WasmKit) is
included in recent development snapshots of the swift.org toolchain, we
should select it as a default debugger and test runner when a Wasm
triple is selected. User toolsets can still override it if needed
through toolset merging algorithm described in
[SE-0387](https://github.com/swiftlang/swift-evolution/blob/main/proposals/0387-cross-compilation-destinations.md).

This allows `swift run` and `swift test` to delegate to WasmKit when
cross-compiling for any Wasm triple.

**Scope**: Limited to cross-compilation.
**Risk**: Low due to limited scope and added test coverage.
**Testing**: Add new automated test cases, manual testing with Wasm
products.
**Issue**: rdar://150382758
**Reviewer**: @kateinoigakukun @jakepetroules