SwiftFixIt: Always rely on the UTF-8 offset of an input source location
As it turns out, contrary to an old comment in the diagnostic
deserializer, the Swift compiler *does* serialize the offset of a
source location, so stop assuming that it may have been left blank. This
slightly simplifies the diagnostic conversion logic and also enables us
to stop hashing the line and column in the diagnostic filter.