Generate the compatibility header for libraries on all platforms (#8695)
Enable generating the compatibility header for library targets on all
platforms by lifting the restriction to Darwin.

As part of the work to formalize `@cdecl` we've been adding content to
the compatibility header that is usable by C compilers without the need
for Objective-C or modules. Lifting this restriction will allow Linux
and Windows users to generate and import the compatibility header to
call `@cdecl` functions in their package.