Leverage package graph, add include paths. (#8750)
Take advantage of the fact we're in swiftpm and calculate things out of
the package graph instead of hard coding. In particular, we compute the
list of targets for the requested list of products, as well as the list
of C modules and their include paths.

Swift build is better at ensuring the include path is added to all
targets that need it. Unfortunately when mixing libraries that use
swift-syntax and macros, we end up adding both the C modules from the
checkouts and the prebuilts causing duplicate definition of the C
modules. By adding the above include paths and leveraging that we always
checkout the source anyway, switch to using the checked out include
paths, thus avoiding the duplicate.

Also some prep for adding a config.json file so the list of prebuilts
can be changed without a PR to SwiftPM. Removes the magic we were using
to add the prebuilt library to the package manifest and just add them at
the end.