Binary Static Library Artifact auditing tool (#8741)
This is an implementation of the auditing tool described in
[SE-0482](https://github.com/swiftlang/swift-evolution/blob/main/proposals/0482-swiftpm-static-library-binary-target-non-apple-platforms.md).

### Motivation:

As part of SE-0482 and
https://github.com/swiftlang/swift-package-manager/pull/8639 we
introduced the ability to depend on prebuilt static libraries (that
expose a C interface) and that don't have any dependencies outside of
the C standard library. This PR introduces an auditing tool that checks
the ABI of static library artifact bundle and checks if it's compatible
with the current host platform.

### Modifications:

- Create a new package subcommand that checks a local artifact bundle
for unexpected external dependencies.
- New internal APIs to inspect the ABI of a binary object (object file,
static archive, dynamic library).

### Result:

Users will be able to validate that their static library binary
artifacts won't cause runtime issues for users.