Prevent unnecessary rebuilds when modifying swiftc arguments
Many arguments for the compiler are marked as
`doesNotAffectIncrementalBuild` in `swift-driver`. Use the argument
hashing in swift-driver to hash the swift compiler flags and incorporate
that in to the build hash signature used to determine if SwiftPM should
do a full rebuild.

This prevents unnecssary rebuilds when adding or changing arguments like
`-Xswiftc -diagnostic-style=llvm`. For a list of arguments that no
longer affect rebuilds, search for `doesNotAffectIncrementalBuild` in
https://github.com/swiftlang/swift-driver/blob/main/Sources/SwiftOptions/Options.swift.

This builds on the work introduced in #8717 which avoided unnecessary
rebuilds when SwiftPM arguments changed.