[Package{Model, Graph}] Make it possible to mark modules as implicit
It's currently impossible to distringuish between modules that
are created from the declarations in a manifest and the ones
synthesized as part of module graph building and/or build plan
creation. This was only a problem for `SystemLibraryModule` before,
but with the introduction of `swift package migrate` command the
situation changed because the command can only operate on modules
that appear in the manifest file because it won't be possible to
update code or settings of the synthesized modules.