Support using a Swift file as a build tool plugin executable
A Swift file can be executed directly in the shell. Typically the file
begins with a `#/usr/bin/swift` or an equivalent `xcrun` incantation,
and then simply running `swift MyFile.swift` will automatically compile
and execute it.

Build tool plugins often leverage Swift, but to do so they add two
targets, a `.plugin` and an `.executableTarget`. The executable target
builds a binary that the plugin tool then calls during a build.

Often this is more heavyweight than needed. By passing the path to a
Swift file to the `executable` argument of the `.buildCommand`
returned from your build tool plugin, you could bypass the need to
create a separate executable target and just execute the file directly.

However, because SwiftPM executes build tool plugins within a sandbox
the sandbox script needs to be ammended with the clang module cache
temporary directory so that the swift compiler can write the modules it
needs to execute the file.

Ammend the sandbox script to add write permissions to the root tmp
directory that LLVM uses to write clang modules. With this addition you
can now use a .swift file as a build tool plugin executable.

rdar://152874736