adds documentation to define the default value for target isolation (#8823)
adds documentation to define the default value for target isolation

### Motivation:

Following up from [my question in the Swift Forums about the default
value for "isolation
mode"](https://forums.swift.org/t/what-is-the-default-isolation-mode-for-swift-packages-6-2/80453)
for Swift packages.
The default is defined in the [proposal
SE-0466](https://github.com/swiftlang/swift-evolution/blob/main/proposals/0466-control-default-actor-isolation.md),
but wasn't reflected in PackageDescription documentation.

### Modifications:

Updated the documentation for `defaultIsolation` to describe the
default.

### Result:

PackageDescription documentation reflects the default choice if
unspecified or set to `nil`