Do not include all of the build dir for API digester to avoid conflicting definitions. (#8776)
Do not include all of the build dir for API digester to avoid
conflicting definitions.

### Motivation:

Issue #8081 : when a package or its dependencies define a tool plugin,
then diagnose-api-breaking-changes report false API breakage.

Investigation shown that diagnose-api-breaking-changes passes for
analysis (with `-I`) all of the build dir, with all of the dependencies.
When a tool plugin is in the dependency tree, then dependencies for the
plugin itself are built with `.host` destination. If it happens that
these modules are also built for `.target` destination, then
`swift-api-digester` fails to analyze them, and public API declared
there is printed as removed.

```
/Users/yy/work/pub/tst/.build/arm64-apple-macosx/debug/MyLib-tool.build/module.modulemap:1:8: error: redefinition of module 'MyLib'
module MyLib {
       ^
/Users/yy/work/pub/tst/.build/arm64-apple-macosx/debug/MyLib.build/module.modulemap:1:8: note: previously defined here
module MyLib {
       ^
Failed to load module: MyLib
```

### Modifications:

The goal for this PR is to reduce analyzed surface from "all of the
build directory" down to modules built for target.

### Result:

Before the change, example in Issue #8081 

```
1 breaking change detected in mypackage:
  💔 API breakage: var myPublicString has been removed
```

After the change:

```
No breaking changes detected in mypackage
```