Migrate PluginTests to Swift Testing (#8793)
Migrate the test suite to Swift Testing.

Explicitly mark tests that are enabled/disabled.

Add bug links to issues and radars for better tracking combined
withKnownIssues so that when bugs are fixed there is immediate
notification that the tests can be enabled.

Fix the writable directories and working directory in PIFBuilder
custom tasks to the plugin working directory for plugins that don’t declare their outputs.