Fix --quiet option not working with swift run #8844
Use the log level to determine whether a message should be log or not. If log level is error, log only error messages and ignore all other messages.
This fixes the issue for swift run command, and it benefits the swift build command as well.

Add a small refactoring to remove duplicated decleration of the isVerbose property from different targets.