Activate swift build support for command plugins (#8845)
Be consistent in the detection of the build system settings so that both
the build system and the build parameters are using the same setting. Use
the build system setting given after the command plugin name as the highest
precedence, followed by the one (if any) provided after the package
subcommand.

When building a specific target with the swiftbuild build system, set
the configured target in the build request to be the one that isn't a
dynamic target.

Tag the `TraitTest.packagePluginGetSymbolGraph_enablesAllTraits` with a
known issue when using the swiftbuild build system since the symbol graph
extraction is not yet working there.

Add the toolchain's runtime environment to the environment used when
running command plugins so that the tools have the necessary runtime
dependencies when the plugin decides to run them on platforms like Linux.