SwiftBuild: Enable traits support (#8873)
When Creasting the SwiftBuildSystem Factory, the traits configuration
was not passed when loading the package graph.

Update the factory to pass the traits configuration during package
resolution, and remove the `withKnownIssues` in TraitTests as it relates
to SwiftBuild.

There are some know issues, which are tracked via the automated tests.