Implement diagnose-api-breaking-changes support with --build-system swiftbuild (#8857)
depends on https://github.com/swiftlang/swift-build/pull/601
closes https://github.com/swiftlang/swift-package-manager/issues/8847

Swift Build has builtin support for generating and comparing to API/ABI
baselines. Wire this up to the diagnose-api-breaking-changes command.
Notably, we now have a mechanism for running a Swift Build based build
and collecting serialized diagnostics, which should be reusable in the
implementation of the migrate command.