Disable check for unsafe flags.
As discussed in the Swift forums below, a number of projects
require the use of "unsafe" flags. A common example is the use
of absolute header include paths set via an environment variable
that project consumers agree to use. As such, the flags can be
safely used under those circumstances.

This change disables the one line that was doing the check.

We will further study the issue and come up with other mechanisms
to ensure the challenges some flags have on quality
of the package ecosystem are mitigated.