Revising DocC content for the CLI commands to be able to read the arguments for the commands (#8883)
Updates to all existing CLI reference documentation to make CLI
arguments readable.

### Motivation:

The initial run of the argument parser generator dumped the arguments
into a single, concatenated string that ended up being pretty unreadable
through DocC. This update fixes that to present the arguments in a
wrapped-width code block, aligned at 60 characters to make it easier to
read in DocC.

This matches the PR being made to swift argument parser to revise the
generation of that content:
https://github.com/apple/swift-argument-parser/pull/791

### Modifications:

Updated the CLI arguments to block formatting for all existing CLI
reference documentation pages.

### Result:

The arguments for any of the CLI documents are much easier to read.