Package: Update Swift Tool Support Core references (#8910)
When the Integration Tests were merged with the Tests (#8223), the
`Package.swift` added dependencies on Swift Tools Support Core. However,
the declaration did not take into account when the
SWIFTPM_SWBUILD_FRAMEWORK environment variable was defined, which was
updated in #8442 and #8802.

Update the declaration of any Swift Tools Support Core dependencies to
use the respective local variables.