Fix `swift package describe` for packages with binary artifacts that has a .zip extension. (#8809)
Fix `swift package describe` for packages with binary artifacts that has
a .zip extension.

### Motivation:

Fixes #4387

I want to avoid an error when running `swift package describe` on a
package that contains a binary artifact target with a `.zip` extension.

### Modifications:

I modified the
`Workspace.loadRootPackage(at:observabilityScope:completion)` method.
Specifically, I updated the logic for local binary targets to also
return a `BinaryArtifact` object when the file extension is `.zip`.
This change is also based on the comment referenced
[here](https://github.com/swiftlang/swift-package-manager/pull/5826#issuecomment-1282990236).

### Result:

When running `swift package describe` on a package that contains a
binary artifact target with a `.zip` extension, the package description
is now output without throwing an error.