Add trait configuration to SwiftCommandState
Since the trait configuration is used for a variety of swift package subcommands,
it makes it a lot easier to add the trait options as a part of the GlobalOptions which
is consumed by SwiftCommandState and persisted across all SwiftCommands.

This also allows us to omit passing a trait configuration to the build system factory
methods, and instead replacing it with a flag that can override the trait configuration
to enable all traits for edge-case scenarios (e.g. fetching symbol graphs)